Fix ConsulBundleTest so it works when a Consul agent is running
Choose a random open port and mock the ConsulBundle#getConsulAgentPort
so that it returns that open port. This guarantees that even if there
is a Consul agent running locally (on any port), this test will pass
because it is trying to connect to an open port.

Added kiwi as a *test* dependency to use its LocalPortChecker. If
we refactor this library to use kiwi in production code, then the
dependency will need to be changed to remove the test scope.

Fixes #100