Update Javadocs in ConsulAdvertiser re: health check URLs
The original design of getHealthCheckUrl in ConsulAdvertiser is
fundamentally broken. The getHealthCheckUrl method was made
protected to allow customization of the health check URL, but no
mechanism was provided to tell ConsulBundle to use a custom
ConsulAdvertiser subclass. Therefore, ConsulBundle always
instantiates ConsulAdvertiser directly, meaning it won't use
any subclass, meaning it can't ever use the custom health check
URL logic.

This commit adds API notes to the original getHealthCheckUrl
method, which is no longer used at all, and the new method. This
is to explain that there is currently no good way to customize the
health check URL, other than hacking it by. For example, by creating
a custom ConsulAdvertiser subclass, creating a custom ConsulBundle
that uses it, and changing application code to use the custom
bundle.