Rename applicationScheme to scheme in ConsulAdvertiser#getHealthCheckUrl
Annoyingly, when I added support for separate application and admin schemes
recently (#269), I did not noticed that the parameter on getHealthCheckUrl was
named applicationScheme, which is no longer correct. So, rename to the more
generic 'scheme'. I also renamed the argument on the deprecated getHealthCheckUrl
method too, so that things are consistent.