Fix test failure caused by consul-client 1.9.0 upgrade
The Consul.Builder#build method no longer eagerly pings Consul
when called. So, we can't force a failure anymore just by telling
it to connect to an open port where nothing is running.

Updated the existing test to validate the new behavior and added
a new test to validate the original behavior (which should not
happen anymore, but the test forces it).